Update testsuite, and fix tests (#90)
* Modify AST parser to disallow any other order than `param`, `result`,
  `local`.
* Allow parsing all var integers as int64 (the lookups will fail later.)
* Fix interpreter logic determining whether to adjust the stack; we were
  correctly handling the case where `unreachable` was being used in an
  expression (the value stack is adjusted "up" as if a value was
  actually produced), but the same behavior is required for `br`,
  `br_table` and `return`.
* Add NAT token to lexer/parser, which is an integer value without a
  sign.

+++ src/wasm-ast.h
@@ -34,7 +34,7 @@ typedef struct WasmVar {
   WasmLocation loc;
   WasmVarType type;
   union {
-    int index;
+    int64_t index;
     WasmStringSlice name;
   };
 } WasmVar;
